refactor: applied the slices pattern to the useGameStore
